Split up Alert Messages into folder
